Run vimboy if the string 'vimboy' is included in the filetype
Vim has a function that allows you to specify multiple filetypes for the same
file (see http://vimdoc.sourceforge.net/htmldoc/options.html#'filetype') by
concatenating them with periods.

I would like to use vimboy and markdown syntax at the same time, so I want
vimboy to work with files of filetype 'markdown.vimboy', which it didn't
before this commit.

This single-character change could probably be done more cleanly, but is
sufficient for my purposes, as it is somewhat unlikely for the string 'vimboy'
to appear in unrelated filetypes.