fix: make sure that vimboy_dir is loaded when defining the syntax
The order of execution of the syntax and ftplugin scripts seems to vary.
This change makes sure, that the buffer variable is set.

Before, we checked `g:vimboy_loaded`.
The new check for `b:vimboy_dir` does also what we want,
because the latter will be set while loading the plugin as well.

fixes #5