fix: open links case insensitive independent of configuration
The main change is switching from `=~` to `=~?`, which matches the
filename case insensitive _independent_ of the users' configuration.
As this loop only is executed when we don't find the exact match, this
is always what should happen.

While at it, remove an unnecesary path divider and improve code
documentation a little bit.